Avance de la decompilacion hasta el 2026-09-18 (foto limpia)
Foto del estado del proyecto en un solo commit, sobre 59f184a0 ("5% linked!
(#123)", 2026-08-21), que es el punto donde esta linea se separo de upstream.
Contiene el trabajo de 1.521 commits locales; el historial fino se conserva en
local y no se sube porque arrastra un archivo que no puede publicarse (un 7z de
108 MB que entro por accidente el 19 de agosto y se quito en el commit
siguiente) y 193 MB de trabajo temporal de rondas.

Fuera de la foto, a proposito: tools/scratch/ (trabajo temporal de rondas).
Dentro no hay ni un binario del juego: orig/ solo lleva sus .gitkeep.

main sigue a upstream; esta rama es la linea de trabajo propia.

Estado medido con scripts/pctall.py (objdiff por version, metodo de upstream),
las cinco versiones con 0 objetos que no compilan:

    GOWE69 (GameCube)          97,14 %  18.298 funciones   DOL 9619ba57...
    SLES-53558-A124 (PS2 EU)   28,41 %  12.230 funciones
    SLUS-21351 (PS2 US)        32,33 %   9.879 funciones
    EUROPEGERMILESTONE (X360)   1,93 %   2.203 funciones
    SPEED_EXE_1_3 (PC)          9,99 %   5.788 funciones

Como se mide y como se trabaja: docs/METODO.md y docs/PROCEDIMIENTO.md.
La puerta de entrada: docs/INDICE.md.

+++ src/Speed/Indep/bWare/Inc/Espresso.hpp
@@ -12,9 +12,9 @@ struct FloatVector {
 };
 
 struct LongVector {
-    float x; // offset 0x0, size 0x4
-    float y; // offset 0x4, size 0x4
-    float z; // offset 0x8, size 0x4
+    long x; // offset 0x0, size 0x4
+    long y; // offset 0x4, size 0x4
+    long z; // offset 0x8, size 0x4
 };
 
 // TODO from debug build
